pyproject.toml
[project]
name = "CESTA"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "seaborn>=0.13.2",
    "torch>=2.0.0,<2.6.0",
    "torch-geometric>=2.7.0",
    "transformers>=5.2.0",
    "typer>=0.15.0",
    "pyyaml>=6.0",
    "dulwich>=1.1.0",
    "optuna>=4.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ruff>=0.4.0",
    "pyright>=1.1.0",
    "pandas-stubs>=3.0.0.260204",
    "nbformat>=5.10.4",
    "pydeps>=3.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/CESTA"]

[project.scripts]
cesta = "CESTA.cli:main"

[tool.ruff]
src = ["src"]
line-length = 150
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["CESTA"]
combine-as-imports = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
